requirements.txt
absl-py==2.1.0
aenum==3.1.15
aiohappyeyeballs==2.4.0
aiohttp==3.11.18
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.11.0
atomicwrites==1.4.1
attrs==24.2.0
autopep8==2.3.2
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==43.0.1
cssselect2==0.8.0
cycler==0.12.1
dataclasses==0.6
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
Elixir==0.7.1
et-xmlfile==1.1.0
fastapi==0.115.12
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.4.1
future==1.0.0
git-filter-repo==2.45.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-documentai==3.5.0
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.63.0
gotrue==2.12.0
greenlet==3.0.3
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.62.3
gspread==6.1.0
gunicorn==23.0.0
h11==0.16.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.0.1
idna==2.10
imap-tools==1.10.0
importlib-metadata==6.11.0
iniconfig==2.0.0
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
line-bot-sdk==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
multidict==6.1.0
narwhals==1.36.0
numpy==1.26.2
oauthlib==3.2.2
openai==1.44.0
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
openpyxl==3.1.3
opt_einsum==3.4.0
packaging==23.2
pandas==2.1.3
pbr==6.1.0
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.1.0
playwright==1.43.0
plotly==6.0.1
pluggy==1.5.0
postgrest==1.0.1
postgrest-py==0.4.0
propcache==0.3.1
proto-plus==1.23.0
protobuf==4.25.3
py==1.11.0
pyarrow==19.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
pydeck==0.9.1
pydyf==0.11.0
pyee==11.1.0
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.1.2
pypdfium2==4.30.0
pyphen==0.17.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pytesseract==0.3.10
pytest==6.2.5
pytest-base-url==2.1.0
pytest-mock==3.14.0
pytest-playwright==0.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2024.1
realtime==2.4.3
realtime-py==0.1.3
referencing==0.36.2
requests==2.25.1
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
scipy==1.15.1
seaborn==0.13.0
secure-smtplib==0.1.1
sentencepiece==0.2.0
shapely==2.1.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.5.1
starlette==0.46.2
storage3==0.11.3
streamlit==1.29.0
StrEnum==0.4.15
supabase==2.15.1
supabase-py==0.0.2
supafunc==0.9.4
tabula-py==2.10.0
Tempita==0.5.2
tenacity==8.5.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.66.5
typing_extensions==4.13.2
tzdata==2024.1
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.2
validators==0.34.0
vertexai==1.71.1
watchdog==6.0.0
weasyprint==65.0
webencodings==0.5.1
websockets==14.2
Werkzeug==3.0.4
wrapt==1.16.0
xlrd==0.7.1
xlwt==0.7.2
yarl==1.20.0
zipp==3.21.0
zopfli==0.2.3.post1
